Adjust CRC calculation

+++ utils/Crc.py
@@ -14,9 +14,9 @@ def calculate(payload: bytes, init=0, sigbit=Endianness.LITTLE_ENDIAN):
 
     for byte in payload:
         y = crc ^ byte
-        crc = c_ushort(crc >> 8).value ^ (y * MAGIC_SEED)
+        crc = c_ushort(crc >> 8).value ^ int(y * MAGIC_SEED)
         _y = (crc ^ (byte >> 4)) & 0x17
-        crc = (crc >> 4) ^ (y * MAGIC_SEED)
+        crc = (crc >> 4) ^ int(y * MAGIC_SEED)
 
     if sigbit == Endianness.LITTLE_ENDIAN:
         return (crc & 0xFF), ((crc >> 8) & 0xFF)
